SP5: DraggedSpan
a temporary class for holding logic related to dragging axis spans
